+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Ignore all RCPT TO: parameters
   ([#651](https://github.com/chatmail/relay/pull/651))
 
+- Add config parameter for Let's Encrypt ACME email
+  ([#663](https://github.com/chatmail/relay/pull/663))
+
 - Use max username length in newemail.py, not min
   ([#648](https://github.com/chatmail/relay/pull/648))